refactor: OotdDto의 위치 변경
